Removed batch comment
Removed a commented line that I had commented out so that I could test the implementation of removing duplicates. With the comment in, it would not convert any pdfs, it would just work off of existing CSV's. Now it will convert the PDFs again. Also added myself as an author lol